requirements.txt
flask
redis
requests
psycopg2-binary
